requirements.txt
streamlit
python-dotenv
langchain
langchain-community
faiss-cpu
groq
langchain-ollama
pymilvus
tqdm
pypdf
bs4
langserve
httpx
fastapi
python-multipart
sse_starlette
uvicorn
aiohttp
asyncio